feat: rerender quiz on "try again"
Rerender the quiz on the "try again" button is
clicked. NOTE that currently the quiz is
rerendered exactly the same. I suggest that
we could pass in a question list that is greater
than the amount of question that we want to
render, then we could shuffle it and select a
subset on rerender.